done with lab2 tasks

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,26 +5,26 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+    public static long divideThenRound(double top, double bottom) {
+        double quotient = top / bottom;
+        long result = Math.round(quotient);
         return result;
     }
 
     public static void main(String[] args) {
         int t = 10;
         int b = 2;
-        int result = divideThenRound(t, b);
+        long result = divideThenRound(t, b);
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
         int t2 = 9;
         int b2 = 4;
-        int result2 = divideThenRound(t2, b2);
+        long result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
         int t3 = 3;
         int b3 = 4;
-        int result3 = divideThenRound(t3, b3);
+        long result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
     }
 }
